Move conversion code to own file
